+++ app/routers/candidates.py
@@ -1,7 +1,7 @@
 from bofire.data_models.dataframes.api import Candidates
 from fastapi import APIRouter
 
-from bofire_candidates_api.api_data_models import CandidatesRequest
+from bofire_candidates_api.data_models import CandidatesRequest
 from bofire_candidates_api.generate import generate_candidates
