+++ bofire_candidates_api/worker.py
@@ -8,7 +8,7 @@ import requests
 from bofire.data_models.dataframes.api import Candidates
 from pydantic import BaseModel, model_validator
 
-from bofire_candidates_api.api_data_models import CandidatesProposal, ProposalStateEnum
+from bofire_candidates_api.data_models import CandidatesProposal, ProposalStateEnum
 from bofire_candidates_api.generate import generate_candidates
