+++ tests/test_worker.py
@@ -5,7 +5,7 @@ from bofire.benchmarks.api import Himmelblau
 from bofire.data_models.dataframes.api import Candidates, Experiments
 from bofire.data_models.strategies.api import RandomStrategy, SoboStrategy
 
-from bofire_candidates_api.api_data_models import CandidatesProposal, CandidatesRequest
+from bofire_candidates_api.data_models import CandidatesProposal, CandidatesRequest
 from bofire_candidates_api.worker import Client as WorkerClient
 from bofire_candidates_api.worker import Worker
 from tests.conftest import Client
